+++ lib/moving-average/CHANGELOG.md
@@ -1,7 +1,8 @@
 <!-- moving-average -->
 
 ## 0.0.2
-* Added class `SMA_T`, simple moving average.
+* Added class `SMA_T`, simple moving average of unsigned integers.
+* Added class `SMA`, simple moving average.
 * Replaced `filter` method with class operators.
 
 ## 0.0.1